+++ access/prismaAccess.py
@@ -45,7 +45,7 @@ class prismaAccess:
 		#self.hipObjectsUri = self.saseApi + __configV1 + "hip-objects"
 		#self.hipProfilesUri = self.saseApi + __configV1 + "hip-profiles"
 		#self.quarantinedDevicesUri = self.saseApi + __configV1 + "quarantined-devices"
-		#self.regionsUri = self.saseApi + __configV1 + "regions"
+		self.regionsUri = self.saseApi + __configV1 + "regions"
 		self.schedulesUri = self.saseApi + __configV1 + "schedules"
 		self.serviceGroupsUri = self.saseApi + __configV1 + "service-groups"
 		self.servicesUri = self.saseApi + __configV1 + "services"
@@ -87,7 +87,7 @@ class prismaAccess:
 
 	def __init__(self, __saseToken):
 		"""Initialize Class"""
-		self.prismaAccessPythonAPIVersion = "0.18"
+		self.prismaAccessPythonAPIVersion = "0.19"
 		self.saseApi = "https://api.sase.paloaltonetworks.com"
 		self.saseToken = __saseToken
 		self.contentType = "application/json"